Convert test_planner to exported library
