+++ CMakeLists.txt
@@ -52,6 +52,30 @@ target_link_libraries(
   PRIVATE
     rmf_traffic::rmf_traffic
     rmf_fleet_adapter::rmf_fleet_adapter
+    rmf_performance_tests
+)
+
+target_include_directories(
+    test_planner
+    PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+)
+
+add_library(rmf_performance_tests SHARED src/rmf_performance_tests.cpp src/Scenario.cpp)
+
+target_link_libraries(
+    rmf_performance_tests
+    PRIVATE
+    rmf_traffic::rmf_traffic
+    rmf_fleet_adapter::rmf_fleet_adapter
+)
+
+target_include_directories(
+    rmf_performance_tests
+    PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
 )
 
 # TODO(MXG): If we ever want to distribute this test application in a debian
@@ -59,6 +83,13 @@ target_link_libraries(
 target_compile_definitions(
   test_planner
   PRIVATE
+    "TEST_MAP_DIR=\"${OUTPUT_MAP_DIR}\""
+    "TEST_SCENARIO_DIR=\"${CMAKE_CURRENT_BINARY_DIR}/scenarios/\""
+)
+
+target_compile_definitions(
+    rmf_performance_tests
+    PRIVATE
     "TEST_MAP_DIR=\"${OUTPUT_MAP_DIR}\""
     "TEST_SCENARIO_DIR=\"${CMAKE_CURRENT_BINARY_DIR}/scenarios/\""
 )
@@ -71,5 +102,26 @@ foreach(scenario ${scenarios})
   configure_file(${scenario} "${CMAKE_CURRENT_BINARY_DIR}/scenarios/${scenario_file}" COPYONLY)
 
 endforeach()
+
+
+install(
+    DIRECTORY include/
+    DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
+)
+
+ament_export_targets(rmf_performance_tests HAS_LIBRARY_TARGET)
+
+install(
+    TARGETS rmf_performance_tests
+    EXPORT rmf_performance_tests
+    RUNTIME DESTINATION lib/rmf_performance_tests
+    LIBRARY DESTINATION lib
+    ARCHIVE DESTINATION lib
+)
+
+install(
+    DIRECTORY include/
+    DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
+)
 
 ament_package()
